+++ language/english_utf-8.php
@@ -12,7 +12,7 @@
  * @category   glFusion CMS
  * @package    glFusion
  * @author     Mark R. Evans  mark AT glFusion DOT org
- * @copyright  2018 - Mark R. Evans
+ * @copyright  2018-2022 - Mark R. Evans
  * @license    http://opensource.org/licenses/gpl-2.0.php - GNU Public License v2 or later
  *
  */
@@ -63,10 +63,10 @@ $LANG_fs['privacy'] = array(
     'pv_cookie'   => 'Consent Cookie Settings',
 );
 
-$LANG_configselects['privacy'] = array(
-    0 => array('True' => 1, 'False' => 0),
-    1 => array('Bottom' => 'bottom', 'Top' => 'top', 'Top (pushdown)' => 'top_pushdown', 'Float Right' => 'bottom-right', 'Float Left' => 'bottom-left'),
-    2 => array('Block' => 'block', 'Classic' => 'classic', 'Edgeless' => 'edgeless'),
+$LANG_configSelect['privacy'] = array(
+    0 => array(1 => 'True', 0 => 'False'),
+    1 => array('bottom' => 'Bottom', 'top' => 'Top', 'top_pushdown' => 'Top (pushdown)', 'bottom-right' => 'Float Right', 'bottom-left' => 'Float Left'),
+    2 => array('block' => 'Block', 'classic' => 'Classic', 'edgeless' => 'Edgeless'),
 );
 
 $PLG_privacy_MESSAGE1 = 'Privacy Plugin upgrade: Update completed successfully.';