+++ upgrade.php
@@ -12,7 +12,7 @@
  * @category   glFusion CMS
  * @package    glFusion
  * @author     Mark R. Evans  mark AT glFusion DOT org
- * @copyright  2018-2021 - Mark R. Evans
+ * @copyright  2018-2022 - Mark R. Evans
  * @license    http://opensource.org/licenses/gpl-2.0.php - GNU Public License v2 or later
  *
  */
@@ -37,6 +37,9 @@ function privacy_upgrade()
             // no changes
 
         case '1.0.2' :
+            // no changes
+
+        case '1.0.3' :
             // no changes
 
         default :